Rename src/error to src/error-reporting
