+++ ads/inabox/inabox-host.js
@@ -22,7 +22,7 @@
 import {InaboxMessagingHost} from './inabox-messaging-host';
 import {dev, initLogConstructor, setReportError, user} from '../../src/log';
 import {getData} from '../../src/event-helper';
-import {reportError} from '../../src/error';
+import {reportError} from '../../src/error-reporting';
 
 /** @const {string} */
 const TAG = 'inabox-host';
