+++ extensions/amp-access/0.1/amp-login-done.js
@@ -24,7 +24,7 @@ import {LoginDoneDialog} from './amp-login-done-dialog';
 import {bodyAlwaysVisible} from '../../../src/style-installer';
 import {initLogConstructor, setReportError} from '../../../src/log';
 import {onDocumentReady} from '../../../src/document-ready';
-import {reportError} from '../../../src/error';
+import {reportError} from '../../../src/error-reporting';
 
 bodyAlwaysVisible(window);
 initLogConstructor();
