+++ extensions/amp-ad/0.1/amp-ad-xorigin-iframe-handler.js
@@ -32,7 +32,7 @@ import {getHtml} from '../../../src/get-html';
 import {isExperimentOn} from '../../../src/experiments';
 import {isGoogleAdsA4AValidEnvironment} from '../../../ads/google/a4a/utils';
 import {removeElement} from '../../../src/dom';
-import {reportErrorToAnalytics} from '../../../src/error';
+import {reportErrorToAnalytics} from '../../../src/error-reporting';
 import {setStyle} from '../../../src/style';
 import {throttle} from '../../../src/utils/rate-limit';
