+++ extensions/amp-lightbox-gallery/0.1/amp-lightbox-gallery.js
@@ -51,7 +51,7 @@ import {getElementServiceForDoc} from '../../../src/element-service';
 import {htmlFor} from '../../../src/static-template';
 import {isExperimentOn} from '../../../src/experiments';
 import {prepareImageAnimation} from '@ampproject/animations';
-import {reportError} from '../../../src/error';
+import {reportError} from '../../../src/error-reporting';
 import {setStyle, setStyles, toggle} from '../../../src/style';
 import {toArray} from '../../../src/core/types/array';
 import {triggerAnalyticsEvent} from '../../../src/analytics';
