+++ src/custom-element.js
@@ -36,7 +36,7 @@ import {
   isBlockedByConsent,
   isCancellation,
   reportError,
-} from './error';
+} from './error-reporting';
 import {dev, devAssert, rethrowAsync, user, userAssert} from './log';
 import {getIntersectionChangeEntry} from './utils/intersection-observer-3p-host';
 import {getMode} from './mode';
