+++ src/service/custom-element-registry.js
@@ -18,7 +18,7 @@ import {ElementStub} from '../element-stub';
 import {Services} from '../services';
 import {createCustomElementClass, stubbedElements} from '../custom-element';
 import {extensionScriptsInNode} from './extension-script';
-import {reportError} from '../error';
+import {reportError} from '../error-reporting';
 import {userAssert} from '../log';
 
 /** @type {!WeakMap<!./service/ampdoc-impl.AmpDoc, boolean>} */
