+++ src/service/real-time-config/real-time-config-impl.js
@@ -19,7 +19,7 @@ import {Services} from '../../services';
 import {dev, user, userAssert} from '../../log';
 import {getMode} from '../../mode';
 import {isArray, isObject} from '../../core/types';
-import {isCancellation} from '../../error';
+import {isCancellation} from '../../error-reporting';
 
 import {registerServiceBuilderForDoc} from '../../service';
 import {tryParseJson} from '../../json';
