+++ src/service/resource.js
@@ -17,7 +17,11 @@
 import {Deferred} from '../utils/promise';
 import {Layout} from '../layout';
 import {Services} from '../services';
-import {cancellation, isBlockedByConsent, reportError} from '../error';
+import {
+  cancellation,
+  isBlockedByConsent,
+  reportError,
+} from '../error-reporting';
 import {computedStyle, toggle} from '../style';
 import {dev, devAssert} from '../log';
 import {
