+++ test/unit/test-resource.js
@@ -20,7 +20,7 @@ import {OwnersImpl} from '../../src/service/owners-impl';
 import {Resource, ResourceState} from '../../src/service/resource';
 import {ResourcesImpl} from '../../src/service/resources-impl';
 import {Services} from '../../src/services';
-import {isCancellation} from '../../src/error';
+import {isCancellation} from '../../src/error-reporting';
 import {layoutRectLtwh} from '../../src/layout-rect';
 
 describes.realWin('Resource', {amp: true}, (env) => {
